1121代码

+++ LoveBank.Web.Admin/Controllers/ProductController.cs
@@ -42,15 +42,32 @@ namespace LoveBank.Web.Admin.Controllers
             {
 
                 var t_p = db.T_Product;
+                var t_d = db.T_Department;
 
-                var list = from p in t_p select p;
+                //var list2 = from p in t_p select p;
+                var list = from p in t_p
+                           join d in t_d on p.DeptId equals d.Id
+                           select new ProductModel
+                           {
+                               Id = p.ID,
+                               Name = p.Name,
+                               DeptId = p.DeptId,
+                               Count = p.Count,
+                               Price = p.Price,
+                               BarCode = p.BarCode,
+                               CostScore = p.CostScore,
+                               EndTime = p.EndTime,
+                               StartTime = p.StartTime,
+                               State = p.State,
+                               DeptIdName = d.Name
+                           };
 
                 list = list.Where(x => x.State != RowState.删除 && x.DeptId.IndexOf(AdminUser.DeptId) > -1);
 
                 if (!string.IsNullOrEmpty(model.Name)) list = list.Where(x => x.Name.Contains(model.Name));
                 if (!string.IsNullOrEmpty(model.BarCode)) list = list.Where(x => x.BarCode == model.BarCode);
 
-                model.ProductList = list.OrderByDescending(x => x.ID).ToPagedList(pageNumber - 1, size);
+                model.ProductList = list.OrderByDescending(x => x.Id).ToPagedList(pageNumber - 1, size);
                 return View(model);
             }
         }
@@ -97,10 +114,25 @@ namespace LoveBank.Web.Admin.Controllers
             model.Price = parm.Price;
             model.BarCode = parm.BarCode;
             model.Type = parm.Type;
-
+            model.Sponsors = parm.Sponsors;
             foreach (var item in parm.SourceFileList)
             {
                 item.Guid = model.Guid;
+                item.AddTime = DateTime.Now;
+
+            }
+
+            model.LogoGuid = Guid.NewGuid().ToString();
+            foreach (var item in parm.SourceFileListLogo)
+            {
+                item.Guid = model.LogoGuid;
+                item.AddTime = DateTime.Now;
+
+            }
+            model.AdGuid = Guid.NewGuid().ToString();
+            foreach (var item in parm.SourceFileListAd)
+            {
+                item.Guid = model.AdGuid;
                 item.AddTime = DateTime.Now;
 
             }
@@ -112,6 +144,8 @@ namespace LoveBank.Web.Admin.Controllers
                 db.Add(model);
                 db.SaveChanges();
                 db.T_SourceFile.AddRange(parm.SourceFileList);
+                db.T_SourceFile.AddRange(parm.SourceFileListLogo);
+                db.T_SourceFile.AddRange(parm.SourceFileListAd);
                 db.SaveChanges();
 
                 return Success("添加成功");
@@ -147,10 +181,13 @@ namespace LoveBank.Web.Admin.Controllers
                                  StartTime = p.StartTime,
                                  State = p.State,
                                  Id = p.ID,
-                                 Desc=p.Desc,
+                                 Desc = p.Desc,
                                  BarCode = p.BarCode,
                                  Type = p.Type,
-                                 SourceFileList = t_s.Where(x => x.Guid == p.Guid).ToList()
+                                 Sponsors = p.Sponsors,
+                                 SourceFileList = t_s.Where(x => x.Guid == p.Guid && !string.IsNullOrEmpty(p.Guid)).ToList(),
+                                 SourceFileListLogo = t_s.Where(x => x.Guid == p.LogoGuid && !string.IsNullOrEmpty(p.LogoGuid)).ToList(),
+                                 SourceFileListAd = t_s.Where(x => x.Guid == p.AdGuid && !string.IsNullOrEmpty(p.AdGuid)).ToList()
                              }).SingleOrDefault();
 
                 return View(model);
@@ -186,7 +223,7 @@ namespace LoveBank.Web.Admin.Controllers
                 model.BarCode = parm.BarCode;
                 model.Desc = parm.Desc;
                 model.Type = parm.Type;
-
+                model.Sponsors = parm.Sponsors;
                 //foreach (var item in parm.SourceFileList)
                 //{
                 //    item.Guid = model.Guid;
@@ -194,7 +231,11 @@ namespace LoveBank.Web.Admin.Controllers
 
                 //}
                 ///删除原来的,彻底以新增方式进行（修改通过删除在新增实现）
-                var delSourceFile = from s in t_s where s.Guid == model.Guid select s;
+                var delSourceFile = from s in t_s
+                                    where (!string.IsNullOrEmpty(model.Guid) && s.Guid == model.Guid)
+                                        || (!string.IsNullOrEmpty(model.LogoGuid) && s.Guid == model.LogoGuid) ||
+                                        (!string.IsNullOrEmpty(model.AdGuid) && s.Guid == model.AdGuid)
+                                    select s;
                 db.T_SourceFile.RemoveRange(delSourceFile);
                 db.SaveChanges();
                 #endregion
@@ -204,11 +245,57 @@ namespace LoveBank.Web.Admin.Controllers
 
                 foreach (var item in parm.SourceFileList)
                 {
-                    item.Guid = model.Guid;
+                    if (string.IsNullOrEmpty(model.Guid))
+                    {
+                        model.Guid = Guid.NewGuid().ToString();
+
+                    }
                     item.AddTime = DateTime.Now;
+                    item.Guid = model.Guid;
+
                 }
 
+
+                foreach (var item in parm.SourceFileListLogo)
+                {
+                    if (string.IsNullOrEmpty(model.LogoGuid))
+                    {
+                        model.LogoGuid = Guid.NewGuid().ToString();
+
+                    }
+                    item.AddTime = DateTime.Now;
+                    item.Guid = model.LogoGuid;
+
+
+                }
+
+                foreach (var item in parm.SourceFileListAd)
+                {
+                    if (string.IsNullOrEmpty(model.AdGuid))
+                    {
+                        model.AdGuid = Guid.NewGuid().ToString();
+
+                    }
+                    item.AddTime = DateTime.Now;
+                    item.Guid = model.AdGuid;
+
+
+                }
                 db.T_SourceFile.AddRange(parm.SourceFileList);//重新绑定
+
+                if (parm.SourceFileListLogo != null &&parm.SourceFileListLogo.Count > 0)
+                {
+                    db.T_SourceFile.AddRange(parm.SourceFileListLogo);
+                }
+
+                if (parm.SourceFileListAd!=null&&parm.SourceFileListAd.Count>0)
+                {
+                    db.T_SourceFile.AddRange(parm.SourceFileListAd);
+                }
+              
+                //db.T_SourceFile.AddRange(parm.SourceFileList);
+
+
                 db.SaveChanges();
 
                 return Success("修改成功");
@@ -233,10 +320,10 @@ namespace LoveBank.Web.Admin.Controllers
             {
                 Error("请选择文件");
             }
-            
-                SourceFile res = UploadFileInstance.SaveFile(file, "ProductImg", AdminUser.ID);
-                return Json(res);
-            
+
+            SourceFile res = UploadFileInstance.SaveFile(file, "ProductImg", AdminUser.ID);
+            return Json(res);
+
         }
 
         /// <summary>
@@ -253,8 +340,8 @@ namespace LoveBank.Web.Admin.Controllers
             var size = pageSize ?? PageSize;
             using (LoveBankDBContext db = new LoveBankDBContext())
             {
-              
-               
+
+
                 var t_mp = db.T_MachineProduct;
                 var t_m = db.T_Machine;
                 var t_d = db.T_Department;
@@ -282,22 +369,22 @@ namespace LoveBank.Web.Admin.Controllers
                            };
 
 
-                list = list.Where(x => x.State != RowState.删除);
+                list = list.Where(x => x.State != RowState.删除 && x.DeptId.IndexOf(AdminUser.DeptId) > -1);
                 if (!string.IsNullOrEmpty(machine.Name)) list = list.Where(x => x.Name.Contains(machine.Name));
-                if (!string.IsNullOrEmpty(machine.MachineCode)) list = list.Where(x => x.MachineCode==machine.MachineCode);
+                if (!string.IsNullOrEmpty(machine.MachineCode)) list = list.Where(x => x.MachineCode == machine.MachineCode);
                 if (!string.IsNullOrEmpty(machine.DeptId)) list = list.Where(x => x.DeptId.IndexOf(machine.DeptId) > -1);
 
 
                 ViewBag.ProductId = machine.ProductId;
                 machine.MachineModelList = list.OrderByDescending(x => x.Id).ToPagedList(pageNumber - 1, size);
 
-                var list2 = t_d.Where(x => x.Level <= 6&&x.Id.IndexOf(AdminUser.DeptId)>-1).ToList();
+                var list2 = t_d.Where(x => x.Level <= 6 && x.Id.IndexOf(AdminUser.DeptId) > -1).ToList();
                 if (!string.IsNullOrEmpty(machine.DeptId))
                 {
                     list2.FirstOrDefault(x => x.Id == machine.DeptId).IsCheck = true;
                     //list2[list2.FindIndex(x => x.Id == machine.DeptId)].IsCheck = true;
                 }
-             
+
 
                 ViewData["Department_List"] = HelpSerializer.JSONSerialize<List<Department>>(list2);
 
@@ -305,34 +392,66 @@ namespace LoveBank.Web.Admin.Controllers
             }
         }
         [SecurityNode(Name = "绑定机器执行")]
-        public ActionResult PostBindMachine(int productId, List<int> machineIdList)
+        public ActionResult PostBindMachine(int productId, List<int> machineIdList, List<int> machineIdInitList)
         {
+
 
             MachineProduct mp;
             List<MachineProduct> machineProductList = new List<MachineProduct>();
-            foreach (var item in machineIdList)
-            {
-                mp = new MachineProduct();
-                mp.AddTime = DateTime.Now;
-                mp.AddUserId = AdminUser.ID;
-                mp.ProductId = productId;
-                mp.MachineId = item;
 
-                machineProductList.Add(mp);
-            }
 
-      
             JsonMessage retJson = new JsonMessage();
             using (LoveBankDBContext db = new LoveBankDBContext())
             {
                 var t_mp = db.T_MachineProduct;
+                var t_m = db.T_Machine;
+
+
+          
+                ///和初始的机器列表对比，找出被删除的机器Id
+                List<int> delli = null;
+                if (machineIdInitList!=null)
+                {
+                    delli = machineIdInitList.FindAll(x => !machineIdList.Contains(x));
+
+                }
+                ////查找出新增的
+                //List<int> addli = machineIdList.FindAll(x => machineIdInitList == null || !machineIdInitList.Contains(x));
+
+                //查找出新增的
+                List<int> addli = null;
+                if (machineIdList != null)
+                {
+                    //查找出新增的
+                    addli = machineIdList.FindAll(x => machineIdInitList == null || !machineIdInitList.Contains(x));
+                }
+
+                if (machineIdList != null && addli != null)
+                {
+                    foreach (var item in addli)
+                    {
+                        mp = new MachineProduct();
+                        mp.AddTime = DateTime.Now;
+                        mp.AddUserId = AdminUser.ID;
+                        mp.ProductId = productId;
+                        mp.MachineId = item;
+
+                        machineProductList.Add(mp);
+                    }
+                }
+
+
 
                 using (TransactionScope transaction = new TransactionScope())
                 {
                     ///删除原来的,彻底以新增方式进行（修改通过删除在新增实现）
-                    var delSourceFile = from s in t_mp where s.ProductId == productId select s;
-                    db.T_MachineProduct.RemoveRange(delSourceFile);
-                    db.SaveChanges();
+                    if (delli!=null&&delli.Count>0)
+                    {
+                        var delSourceFile = from s in t_mp where s.MachineId.HasValue && delli.Contains(s.MachineId.Value) && s.ProductId == productId select s;
+                        db.T_MachineProduct.RemoveRange(delSourceFile);
+                        db.SaveChanges();
+                    }
+                  
 
                     if (machineProductList == null)//null 标示解除了全部绑定
                     {
